+++ client/base/Api_protocol_query.cpp
@@ -246,7 +246,7 @@ bool Api_protocol::parseQuery(const uint8_t &packetCode,const uint8_t &queryNumb
         parseError(QStringLiteral("Procotol wrong or corrupted"),QStringLiteral("remaining data: parseFullQuery(%1,%2 %3) line %4")
                       .arg(packetCode)
                       .arg(QString(data.mid(0,static_cast<int>(in.device()->pos())).toHex()))
-                      .arg(QString(data.mid(static_cast<int>(in.device()->pos()),(in.device()->size()-in.device()->pos())).toHex()))
+                      .arg(QString(data.mid(static_cast<int>(in.device()->pos()),static_cast<int>(in.device()->size()-in.device()->pos())).toHex()))
                       .arg(QStringLiteral("%1:%2").arg(__FILE__).arg(__LINE__))
                       );
         return false;
